lets_train_backend/requirements.txt
Django==1.11.12
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
